sprint5: Next.js 마이그레이션 및 개발 환경 세팅
- auth.js 재작성
- svgr 설치 및 세팅
- 프로젝트 재설정
