--- privacy/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Privacy</title>
-</head>
-<body>
-    <p>Privacy 페이지</p>
-</body>
-</html>
