--- signup/index.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>회원가입 페이지</title>
-    <link rel="stylesheet" href="../styles/globals.css">
-    <link rel="stylesheet" href="../styles/auth.css">
-    <script type="module" src="../scripts/auth.js" defer></script>
-</head>
-<body>
-    <div class="auth-container">
-        <a class="logo-link" href="../index.html">
-            <img class="logo" src="../images/logo.png" alt="logo and link to home page">
-        </a>
-        <div class="auths">
-            <form id="authForm" class="form-container" action="#" method="post">
-                <div class="input-container">
-                    <label for="email">이메일</label>
-                    <input type="email" id="email" name="email" placeholder="이메일을 입력하세요" required>
-                    <div class="error-message" id="email-error"></div>
-                </div>
-                <div class="input-container">
-                    <label for="nickname">닉네임</label>
-                    <input type="text" id="nickname" name="nickname" placeholder="닉네임을 입력하세요" required>
-                    <div class="error-message" id="nickname-error"></div>
-                </div>
-                <div class="input-container">
-                    <label for="password">비밀번호</label>
-                    <div class="password-input-container">
-                        <input type="password" id="password" name="password" placeholder="비밀번호를 입력하세요" required>
-                        <img id="toggle-password" class="icon-button visibility" src="../images/ic_visibility_off.png" alt="숨김 아이콘">
-                    </div>
-                    <div class="error-message" id="password-error"></div>
-                </div>
-                <div class="input-container">
-                    <label for="confirm-password">비밀번호 확인</label>
-                    <div class="password-input-container">
-                        <input type="password" id="confirm-password" name="confirm-password" placeholder="비밀번호를 다시 입력하세요" required>
-                        <img id="toggle-confirm" class="icon-button visibility" src="../images/ic_visibility_off.png" alt="숨김 아이콘">
-                    </div>
-                    <div class="error-message" id="confirm-password-error"></div>
-                </div>
-                <button id="submit-btn" type="submit" class="button submit">회원가입</button>
-            </form>
-            <div class="easy-login">
-                <span>간편 로그인하기</span>
-                <div class="social-icon-container">
-                    <a href="https://www.google.com/" _target="_blank">
-                        <img class="google" src="../images/ic_google.png" alt="구글 로그인 아이콘">
-                    </a>
-                    <a href="https://www.kakaocorp.com/" _target="_blank">
-                        <img class="kakao" src="../images/ic_kakao.png" alt="카카오 로그인 아이콘">
-                    </a>
-                </div>
-            </div>
-            <div class="auth-toggle">
-                <span>이미 회원이신가요?</span>
-                <a class="login-link" href="../login/">로그인</a>
-        </div>
-    </div>
-</body>
-</html>
